% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{countDrug}
\alias{countDrug}
\title{countDrug}
\usage{
countDrug(connectionDetails, cdmSchema, drugName)
}
\arguments{
\item{connectionDetails}{(`list()`)\cr
connectionDetails object created by \link[DatabaseConnector]{createConnectionDetails}.}

\item{cdmSchema}{(`character(1)`)\cr
The name of the cdm schema.}

\item{drugName}{(`character(1)`)\cr
Concept name of a drug.}
}
\value{
(`data.frame()`)
Returns a data frame with the count of the drug.
}
\description{
Counts the amount of records of a specified drug.
}
\examples{
connectionDetails <- Eunomia::getEunomiaConnectionDetails()

countDrug(connectionDetails, "main", "celecoxib")
}
