% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{countDrugCondition}
\alias{countDrugCondition}
\title{countDrugCondition}
\usage{
countDrugCondition(connectionDetails, cdmSchema, drugName, conditionId)
}
\arguments{
\item{connectionDetails}{(`list()`)\cr
connectionDetails object created by \link[DatabaseConnector]{createConnectionDetails}.}

\item{cdmSchema}{(`character(1)`)\cr
The name of the cdm schema.}

\item{drugName}{(`character(1)`)\cr
Concept name of a drug.}

\item{conditionId}{(`numeric(1)`)\cr
Concept ID of a condition.}
}
\value{
(`data.frame()`)
Returns a data frame with the count of the drug within the bounds of the
specified condition.
}
\description{
Counts the amount of records of a specified drug, within the bounds of a
specified condition.
}
\examples{
connectionDetails <- Eunomia::getEunomiaConnectionDetails()

countDrugCondition(connectionDetails, "main", drugName = "celecoxib", conditionId = 192671)
}
