% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{countPersons}
\alias{countPersons}
\title{countPersons}
\usage{
countPersons(connectionDetails, cdmSchema)
}
\arguments{
\item{connectionDetails}{(`list()`)\cr
connectionDetails object created by \link[DatabaseConnector]{createConnectionDetails}.}

\item{cdmSchema}{(`character(1)`)\cr
The name of the cdm schema.}
}
\value{
(`data.frame()`)
Returns a data frame with the count of persons.
}
\description{
Counts the number of persons present in the person table of the OMOP CDM.
}
\examples{
connectionDetails <- Eunomia::getEunomiaConnectionDetails()

countPersons(connectionDetails, "main")
}
